Enhanced ProcessResult, introduce fileName into program arguments, extract chunksize from importer to strategy, fix parallel logic in sizeProviders.

+++ data-pump/src/main/java/com/smikhalev/sqlserverutils/exportdata/TableSizeProvider.java
@@ -3,6 +3,6 @@ package com.smikhalev.sqlserverutils.exportdata;
 import com.smikhalev.sqlserverutils.schema.dbobjects.Table;
 
 public interface TableSizeProvider {
-    public long getSize(Table table);
+    public int getSize(Table table);
     public long getDatabaseSize();
 }
